core: split up `CSSUsage` gatherer
