+++ core/test/audits/byte-efficiency/unused-css-rules-test.js
@@ -30,11 +30,12 @@ describe('Best Practices: unused css rules audit', () => {
 
     const context = {computedCache: new Map()};
 
-    function getArtifacts({CSSUsage, networkRecords = defaultNetworkRecords}) {
+    function getArtifacts({rules, stylesheets, networkRecords = defaultNetworkRecords}) {
       return {
         devtoolsLogs: {defaultPass: networkRecordsToDevtoolsLog(networkRecords)},
         URL: {finalDisplayedUrl: ''},
-        CSSUsage,
+        CSSUsage: rules,
+        Stylesheets: stylesheets,
       };
     }
 
@@ -44,7 +45,7 @@ describe('Best Practices: unused css rules audit', () => {
 
     it('ignores missing stylesheets', () => {
       return UnusedCSSAudit.audit_(getArtifacts({
-        CSSUsage: {rules: [{styleSheetId: 'a', used: false}], stylesheets: []},
+        rules: [{styleSheetId: 'a', used: false}], stylesheets: [],
       }), defaultNetworkRecords, context).then(result => {
         assert.equal(result.items.length, 0);
       });
@@ -52,11 +53,12 @@ describe('Best Practices: unused css rules audit', () => {
 
     it('ignores stylesheets that are 100% used', () => {
       return UnusedCSSAudit.audit_(getArtifacts({
-        CSSUsage: {rules: [
+        rules: [
           {styleSheetId: 'a', used: true},
           {styleSheetId: 'a', used: true},
           {styleSheetId: 'b', used: true},
-        ], stylesheets: [
+        ],
+        stylesheets: [
           {
             header: {styleSheetId: 'a', sourceURL: 'file://a.css'},
             content: '.my.selector {color: #ccc;}\n a {color: #fff}',
@@ -65,7 +67,7 @@ describe('Best Practices: unused css rules audit', () => {
             header: {styleSheetId: 'b', sourceURL: 'file://b.css'},
             content: '.my.favorite.selector { rule: content; }',
           },
-        ]},
+        ],
       }), defaultNetworkRecords, context).then(result => {
         assert.equal(result.items.length, 0);
       });
@@ -73,10 +75,11 @@ describe('Best Practices: unused css rules audit', () => {
 
     it('fails when lots of rules are unused', () => {
       return UnusedCSSAudit.audit_(getArtifacts({
-        CSSUsage: {rules: [
+        rules: [
           {styleSheetId: 'a', used: true, startOffset: 0, endOffset: 11}, // 44 * 25% = 11
           {styleSheetId: 'b', used: true, startOffset: 0, endOffset: 60000}, // 40000 * 3 * 50% = 60000
-        ], stylesheets: [
+        ],
+        stylesheets: [
           {
             header: {styleSheetId: 'a', sourceURL: 'file://a.css'},
             content: '.my.selector {color: #ccc;}\n a {color: #fff}',
@@ -89,7 +92,7 @@ describe('Best Practices: unused css rules audit', () => {
             header: {styleSheetId: 'c', sourceURL: ''},
             content: `${generate('123', 450)}`, // will be filtered out
           },
-        ]},
+        ],
       }), defaultNetworkRecords, context).then(result => {
         assert.equal(result.items.length, 2);
         assert.equal(result.items[0].totalBytes, 100 * 1024);
@@ -116,16 +119,15 @@ describe('Best Practices: unused css rules audit', () => {
         },
       ];
       const artifacts = getArtifacts({
-        CSSUsage: {
-          rules: [
-            {styleSheetId: 'a', used: true, startOffset: 0, endOffset: 60000}, // 40000 * 3 * 50% = 60000
-          ], stylesheets: [
-            {
-              header: {styleSheetId: 'a', sourceURL: 'file://a.html'},
-              content: `${generate('123', 40000)}`, // stylesheet size of 40000 * 3 uncompressed bytes
-            },
-          ],
-        },
+        rules: [
+          {styleSheetId: 'a', used: true, startOffset: 0, endOffset: 60000}, // 40000 * 3 * 50% = 60000
+        ],
+        stylesheets: [
+          {
+            header: {styleSheetId: 'a', sourceURL: 'file://a.html'},
+            content: `${generate('123', 40000)}`, // stylesheet size of 40000 * 3 uncompressed bytes
+          },
+        ],
         networkRecords,
       });
 
@@ -137,10 +139,11 @@ describe('Best Practices: unused css rules audit', () => {
 
     it('does not include empty or small sheets', () => {
       return UnusedCSSAudit.audit_(getArtifacts({
-        CSSUsage: {rules: [
+        rules: [
           {styleSheetId: 'a', used: true, startOffset: 0, endOffset: 8000}, // 4000 * 3 / 2
           {styleSheetId: 'b', used: true, startOffset: 0, endOffset: 500}, // 500 * 3 / 3
-        ], stylesheets: [
+        ],
+        stylesheets: [
           {
             header: {styleSheetId: 'a', sourceURL: 'file://a.css'},
             content: `${generate('123', 4000)}`,
@@ -161,7 +164,7 @@ describe('Best Practices: unused css rules audit', () => {
             header: {styleSheetId: 'e', sourceURL: 'file://e.css'},
             content: '       ',
           },
-        ]},
+        ],
       }), defaultNetworkRecords, context).then(result => {
         assert.equal(result.items.length, 1);
         assert.equal(Math.floor(result.items[0].wastedPercent), 33);
